Add support for .env files in rushx and rush-pnpm.
